Address review findings
- Fix tsconfig.json include to match renamed vite configs
- Narrow bucketInterval to "hour" | "day" union type
- Add trust proxy on root Express app
- Extract SPA fallback into shared helper
- Extract TulipCard into its own component file
- Add error handling to tulips prices route
- Lazy-init Vite dev servers on first subdomain request
- Fix swapped reduce parameter names in eggnet
- Parallel vite builds
- Add error state to tulips frontend
- Share RANGES config between server and client
