+++ packages/oaf/src/server/apps/calendar/app.ts
@@ -1,9 +1,8 @@
 import cors from "cors";
 import express, { type Express } from "express";
-import fs from "node:fs";
-import path from "node:path";
 import type { ViteDevServer } from "vite";
 
+import { spaFallback } from "../../spaFallback.js";
 import { calendarRouter } from "./routes/calendar.js";
 
 export function createCalendarApp(
@@ -21,21 +20,7 @@ export function createCalendarApp(
     .get("/favicon.ico", (_req, res) => void res.send())
     .use("/api/calendar", calendarRouter)
     // SPA fallback
-    .get("*path", async (req, res) => {
-      if (viteDevServer) {
-        const html = fs.readFileSync(
-          path.resolve("src/server/apps/calendar/web/index.html"),
-          "utf-8",
-        );
-        const transformed = await viteDevServer.transformIndexHtml(
-          req.url,
-          html,
-        );
-        res.set("Content-Type", "text/html").send(transformed);
-      } else {
-        res.sendFile(path.resolve("build/client/calendar/index.html"));
-      }
-    });
+    .get("*path", spaFallback(viteDevServer, "calendar"));
 
   return app;
 }
