+++ packages/oaf/src/server/apps/oaf/app.ts
@@ -2,10 +2,9 @@ import bodyParser from "body-parser";
 import cookieParser from "cookie-parser";
 import cors from "cors";
 import express, { type Express } from "express";
-import fs from "node:fs";
-import path from "node:path";
 import type { ViteDevServer } from "vite";
 
+import { spaFallback } from "../../spaFallback.js";
 import {
   authRouter,
   loginHandler,
@@ -59,21 +58,7 @@ export function createOafApp(viteDevServer: ViteDevServer | null): Express {
     .use("/api/resources/message", requireAuth, messageRouter)
     .use("/api/resources/user", requireAuth, userRouter)
     // SPA fallback
-    .get("*path", async (req, res) => {
-      if (viteDevServer) {
-        const html = fs.readFileSync(
-          path.resolve("src/server/apps/oaf/web/index.html"),
-          "utf-8",
-        );
-        const transformed = await viteDevServer.transformIndexHtml(
-          req.url,
-          html,
-        );
-        res.set("Content-Type", "text/html").send(transformed);
-      } else {
-        res.sendFile(path.resolve("build/client/oaf/index.html"));
-      }
-    });
+    .get("*path", spaFallback(viteDevServer, "oaf"));
 
   return app;
 }
