+++ packages/oaf/tsconfig.json
@@ -22,7 +22,7 @@
   "include": [
     "src/**/*.ts",
     "src/**/*.tsx",
-    "vite.config.ts",
+    "vite.*.config.ts",
     "migrations/**/*.ts",
     "kysely.config.ts"
   ],
